feat: move metadata api
