+++ packages/indexer/src/metadata/metadata-api.ts
@@ -19,35 +19,34 @@ export class MetadataApi {
   ): Promise<CollectionMetadata> {
     if (config.liquidityOnly) {
       return await MetadataProvidersMap["onchain"].getCollectionMetadata(contract, tokenId);
-    } else {
-      // get custom / extended metadata locally
-      if (hasCustomHandler(config.chainId, contract)) {
-        const result = await customHandleContractTokens(config.chainId, contract, tokenId);
-        return result;
-      }
-
-      let indexingMethod =
-        options?.indexingMethod ?? MetadataApi.getCollectionIndexingMethod(community);
-
-      //TODO: Remove when adding proper support for overriding indexing method
-      if (config.chainId === 1 && contract === "0xd532b88607b1877fe20c181cba2550e3bbd6b31c") {
-        indexingMethod = "simplehash";
-      }
-
-      if (config.chainId === 137 && contract === "0x2953399124f0cbb46d2cbacd8a89cf0599974963") {
-        indexingMethod = "simplehash";
-      }
-
-      const collection: CollectionMetadata = await MetadataProvidersMap[
-        indexingMethod
-      ].getCollectionMetadata(contract, tokenId);
-
-      if (collection.isFallback && !options?.allowFallback) {
-        throw new Error("Fallback collection data not acceptable");
-      }
-
-      return collection;
     }
+    // get custom / extended metadata locally
+    if (hasCustomHandler(config.chainId, contract)) {
+      const result = await customHandleContractTokens(config.chainId, contract, tokenId);
+      return result;
+    }
+
+    let indexingMethod =
+      options?.indexingMethod ?? MetadataApi.getCollectionIndexingMethod(community);
+
+    //TODO: Remove when adding proper support for overriding indexing method
+    if (config.chainId === 1 && contract === "0xd532b88607b1877fe20c181cba2550e3bbd6b31c") {
+      indexingMethod = "simplehash";
+    }
+
+    if (config.chainId === 137 && contract === "0x2953399124f0cbb46d2cbacd8a89cf0599974963") {
+      indexingMethod = "simplehash";
+    }
+
+    const collection: CollectionMetadata = await MetadataProvidersMap[
+      indexingMethod
+    ].getCollectionMetadata(contract, tokenId);
+
+    if (collection.isFallback && !options?.allowFallback) {
+      throw new Error("Fallback collection data not acceptable");
+    }
+
+    return collection;
   }
 
   public static async getTokensMetadata(
